% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickHeat.R
\name{QuickHeat}
\alias{QuickHeat}
\title{Simple heatmap for GenePix}
\usage{
QuickHeat(dataset, useRank = c("norank", "all", "sample"),
  coltypes = c("darkolivegreen", "darkred"), col, ttl = "")
}
\arguments{
\item{dataset}{GenePix data.}

\item{useRank}{specify if use rank matrix to draw heatmap.
must be one of \code{"norank"} (use original martix, default),
\code{"all"} or \code{"sample"}.}

\item{col}{choose the column you want to look at.}

\item{ttl}{specific title.}

\item{coltype}{color refelcts labeling, green corresponds to IgG; red corresponds to IgM.}

\item{block}{choose the block you want to look at closer.}
}
\description{
A heat map for GenePix Arrays
}

