% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Slide1_IgG}
\alias{Slide1_IgG}
\title{A example of GenePix datasets}
\format{A data frame with 3136 rows and 40 variables:
\describe{
  \item{Block}{block}
  \item{Column}{column}
  ...
}}
\source{
\url{https://microarray.swmed.edu/products/category/protein-array/}
}
\usage{
Slide1_IgG
}
\description{
Four datasets containing the output from GenePix
}
\examples{
data("Slide1_IgG")
data("Slide1_IgM")
data("Slide2_IgG")
data("Slide2_IgM")

}
\keyword{datasets}

