% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlot}
\alias{QuickPlot}
\title{Plot Raw Chip for GenePix}
\usage{
QuickPlot(dataset, coltype, col, ttl = "")
}
\arguments{
\item{dataset}{GenePix data.}

\item{coltype}{color refelcts labeling, green corresponds to IgG; red corresponds to IgM.}

\item{col}{choose the column you want to look at.}

\item{ttl}{specific title.}

\item{block}{choose the block you want to look at closer.}
}
\description{
This function plot GenePix Arrays
}

